Attempt to do something about #404

+++ Common/CUDA/Siddon_projection.cu
@@ -114,20 +114,20 @@ __global__ void kernelPixelDetector( Geometry geo,
         cudaTextureObject_t tex){
     
     
-    unsigned long u = blockIdx.x * blockDim.x + threadIdx.x;
-    unsigned long v = blockIdx.y * blockDim.y + threadIdx.y;
-    unsigned long projNumber=threadIdx.z;
+    unsigned long long u = blockIdx.x * blockDim.x + threadIdx.x;
+    unsigned long long v = blockIdx.y * blockDim.y + threadIdx.y;
+    unsigned long long projNumber=threadIdx.z;
     
     
     if (u>= geo.nDetecU || v>= geo.nDetecV || projNumber>=PROJ_PER_BLOCK)
         return;
     
 #if IS_FOR_MATLAB_TIGRE
-    size_t idx =  (size_t)(u * geo.nDetecV + v)+ (size_t)projNumber*geo.nDetecV *geo.nDetecU ;
+    size_t idx =  (size_t)(u * (unsigned long long)geo.nDetecV + v)+ projNumber*(unsigned long long)geo.nDetecV *(unsigned long long)geo.nDetecU ;
 #else
-    size_t idx =  (size_t)(v * geo.nDetecU + u)+ (size_t)projNumber*geo.nDetecV *geo.nDetecU ;
+    size_t idx =  (size_t)(v * (unsigned long long)geo.nDetecU + u)+ projNumber*(unsigned long long)geo.nDetecV *(unsigned long long)geo.nDetecU ;
 #endif
-    int indAlpha = currProjSetNumber*PROJ_PER_BLOCK+projNumber;  // This is the ABSOLUTE projection number in the projection array (for a given GPU)
+    unsigned long indAlpha = currProjSetNumber*PROJ_PER_BLOCK+projNumber;  // This is the ABSOLUTE projection number in the projection array (for a given GPU)
 
     if(indAlpha>=totalNoOfProjections)
         return;
@@ -138,8 +138,8 @@ __global__ void kernelPixelDetector( Geometry geo,
     Point3D source = projParamsArrayDev[4*projNumber+3];
     
     /////// Get coordinates XYZ of pixel UV
-    int pixelV = geo.nDetecV-v-1;
-    int pixelU = u;
+    unsigned long pixelV = geo.nDetecV-v-1;
+    unsigned long pixelU = u;
     Point3D pixel1D;
     pixel1D.x=(uvOrigin.x+pixelU*deltaU.x+pixelV*deltaV.x);
     pixel1D.y=(uvOrigin.y+pixelU*deltaU.y+pixelV*deltaV.y);
@@ -220,11 +220,11 @@ __global__ void kernelPixelDetector( Geometry geo,
        
     
     // get index of first intersection. eq (26) and (19)
-    int i,j,k;
+    unsigned long i,j,k;
     float aminc=fminf(fminf(ax,ay),az);
-    i=(int)floorf(source.x+ (aminc+am)*0.5f*ray.x);
-    j=(int)floorf(source.y+ (aminc+am)*0.5f*ray.y);
-    k=(int)floorf(source.z+ (aminc+am)*0.5f*ray.z);
+    i=(unsigned long)floorf(source.x+ (aminc+am)*0.5f*ray.x);
+    j=(unsigned long)floorf(source.y+ (aminc+am)*0.5f*ray.y);
+    k=(unsigned long)floorf(source.z+ (aminc+am)*0.5f*ray.z);
     // Initialize
     float ac=am;
     //eq (28), unit anlges
@@ -240,12 +240,12 @@ __global__ void kernelPixelDetector( Geometry geo,
     
     float maxlength=__fsqrt_rd(ray.x*ray.x*geo.dVoxelX*geo.dVoxelX+ray.y*ray.y*geo.dVoxelY*geo.dVoxelY+ray.z*ray.z*geo.dVoxelZ*geo.dVoxelZ);
     float sum=0.0f;
-    unsigned int Np=(imax-imin+1)+(jmax-jmin+1)+(kmax-kmin+1); // Number of intersections
+    unsigned long Np=(imax-imin+1)+(jmax-jmin+1)+(kmax-kmin+1); // Number of intersections
     // Go iterating over the line, intersection by intersection. If double point, no worries, 0 will be computed
     i+=0.5f;
     j+=0.5f;
     k+=0.5f;
-    for (unsigned int ii=0;ii<Np;ii++){
+    for (unsigned long ii=0;ii<Np;ii++){
         if (ax==aminc){
             sum+=(ax-ac)*tex3D<float>(tex, i, j, k);
             i=i+iu;
