+++ Common/CUDA/Siddon_projection_parallel.cu
@@ -107,20 +107,20 @@ __constant__ Point3D projParamsArrayDev[4*PROJ_PER_BLOCK];  // Dev means it is o
 __global__ void kernelPixelDetector_parallel( Geometry geo,
         float* detector, const int currProjSetNumber, const int totalNoOfProjections, cudaTextureObject_t tex){
     
-    unsigned long u = blockIdx.x * blockDim.x + threadIdx.x;
-    unsigned long v = blockIdx.y * blockDim.y + threadIdx.y;
-    unsigned long projNumber=threadIdx.z;
+    unsigned long long u = blockIdx.x * blockDim.x + threadIdx.x;
+    unsigned long long v = blockIdx.y * blockDim.y + threadIdx.y;
+    unsigned long long projNumber=threadIdx.z;
             
     if (u>= geo.nDetecU || v>= geo.nDetecV || projNumber>=PROJ_PER_BLOCK)
         return;
     
-    int indAlpha = currProjSetNumber*PROJ_PER_BLOCK+projNumber;  // This is the ABSOLUTE projection number in the projection array
+    unsigned long indAlpha = currProjSetNumber*PROJ_PER_BLOCK+projNumber;  // This is the ABSOLUTE projection number in the projection array
     
     
 #if IS_FOR_MATLAB_TIGRE
-    size_t idx =  (size_t)(u  * geo.nDetecV + v)+ (size_t)projNumber*geo.nDetecV *geo.nDetecU ;
+    size_t idx =  (size_t)(u  * (unsigned long long)geo.nDetecV + v)+ projNumber*(unsigned long long)geo.nDetecV *(unsigned long long)geo.nDetecU ;
 #else
-    size_t idx =  (size_t)(v  * geo.nDetecU + u)+ (size_t)projNumber*geo.nDetecV *geo.nDetecU ;
+    size_t idx =  (size_t)(v  * (unsigned long long)geo.nDetecU + u)+ projNumber*(unsigned long long)geo.nDetecV *(unsigned long long)geo.nDetecU ;
 #endif
     
     if(indAlpha>=totalNoOfProjections)
@@ -133,8 +133,8 @@ __global__ void kernelPixelDetector_parallel( Geometry geo,
     
 
     /////// Get coordinates XYZ of pixel UV
-    int pixelV = geo.nDetecV-v-1;
-    int pixelU = u;
+    unsigned long pixelV = geo.nDetecV-v-1;
+    unsigned long pixelU = u;
     Point3D pixel1D;
     pixel1D.x=(uvOrigin.x+pixelU*deltaU.x+pixelV*deltaV.x);
     pixel1D.y=(uvOrigin.y+pixelU*deltaU.y+pixelV*deltaV.y);
@@ -220,11 +220,11 @@ __global__ void kernelPixelDetector_parallel( Geometry geo,
     
     
     // get index of first intersection. eq (26) and (19)
-    int i,j,k;
+    unsigned long i,j,k;
     float aminc=fminf(ax,ay);
-    i=(int)floorf(source.x+ (aminc+am)/2*ray.x);
-    j=(int)floorf(source.y+ (aminc+am)/2*ray.y);
-    k=(int)floorf(source.z+ (aminc+am)/2*ray.z);
+    i=(unsigned long)floorf(source.x+ (aminc+am)/2*ray.x);
+    j=(unsigned long)floorf(source.y+ (aminc+am)/2*ray.y);
+    k=(unsigned long)floorf(source.z+ (aminc+am)/2*ray.z);
 //     k=(int)source.z;
     // Initialize
     float ac=am;
@@ -241,12 +241,12 @@ __global__ void kernelPixelDetector_parallel( Geometry geo,
     
     float maxlength=sqrtf(ray.x*ray.x*geo.dVoxelX*geo.dVoxelX+ray.y*ray.y*geo.dVoxelY*geo.dVoxelY);//+ray.z*ray.z*geo.dVoxelZ*geo.dVoxelZ);
     float sum=0.0f;
-    unsigned int Np=(imax-imin+1)+(jmax-jmin+1);//+(kmax-kmin+1); // Number of intersections
+    unsigned long Np=(imax-imin+1)+(jmax-jmin+1);//+(kmax-kmin+1); // Number of intersections
     // Go iterating over the line, intersection by intersection. If double point, no worries, 0 will be computed
     i+=0.5f;
     j+=0.5f;
     k+=0.5f;
-    for (unsigned int ii=0;ii<Np;ii++){
+    for (unsigned long ii=0;ii<Np;ii++){
         if (ax==aminc){
             sum+=(ax-ac)*tex3D<float>(tex, i, j, k);//(ax-ac)*
             i=i+iu;
